LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/mathml/relations/html5-tree/dynamic-childlist-002-expected.txt


FAIL Appending and removing children to maction assert_approx_equals: block position (child 0) expected -26 +/- 1 but got -70
PASS Appending and removing children to menclose
PASS Appending and removing children to merror
FAIL Appending and removing children to mfrac assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 2.375
PASS Appending and removing children to mi
FAIL Appending and removing children to mmultiscripts assert_approx_equals: inline size expected 0 +/- 1 but got 52.75
PASS Appending and removing children to mn
PASS Appending and removing children to mo
FAIL Appending and removing children to mover assert_approx_equals: inline size expected 10 +/- 1 but got 0
PASS Appending and removing children to mpadded
PASS Appending and removing children to mphantom
FAIL Appending and removing children to mroot assert_approx_equals: inline position (child 0) expected 0 +/- 1 but got 4.4375
PASS Appending and removing children to mrow
PASS Appending and removing children to ms
PASS Appending and removing children to mspace
FAIL Appending and removing children to msqrt assert_approx_equals: block size expected 61.78125 +/- 1 but got 81.765625
PASS Appending and removing children to mstyle
FAIL Appending and removing children to msub assert_approx_equals: block position (child 0) expected 0 +/- 1 but got -5.75
FAIL Appending and removing children to msubsup assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 3.1875
FAIL Appending and removing children to msup assert_approx_equals: inline size expected 13.1875 +/- 1 but got 0
PASS Appending and removing children to mtable
PASS Appending and removing children to mtext
FAIL Appending and removing children to munder assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 2.5
FAIL Appending and removing children to munderover assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 6
PASS Appending and removing children to semantics
maction:
mfrac:
mmultiscripts:
mover:
mroot:
msqrt:
msub:
msubsup:
msup:
munder:
munderover:
